HBM-backend/auth/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
pymysql==1.1.0
python-dotenv>=1.0.0
